+++ app/src/main/assets/CHANGES.md
@@ -1,5 +1,9 @@
 # NWS Weather Alerts Widget Change History
 
+## Unreleased Changes
+
+* Fixes a crash at boot on Android 14+
+
 ## Version 2.2.2 (Dec 21, 2025)
 
 * Use a bold-face font for alert titles on the widget to make them easier to read.